refactor(backend): extract per-agent task listing to services/tasks/agent-projection.ts (U2)
Part of the tasks.ts mega-file split (CQ-H1). Moves listTasksByAgent,
projectAgentTaskRows, AGENT_TASK_ACTIVE_STATUSES, and the
AgentTaskActiveStatus type to a sibling submodule.

services/tasks.ts: 884 -> 807 LOC. U3 (zap, ~57 LOC) brings it under
800.

Re-exports preserve the public surface; AgentTaskSummary moves with the
projection function, no longer needed in tasks.ts (also drops the now-
unused inArray drizzle import).

Validation: just check green, 46/46 agents-service + dashboard-agents
route tests pass.
